requirements.txt
streamlit==1.28.1
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.8.5
python-dotenv==1.0.0
